WARNING: MAJOR CHANGE TO INFASTUCTURE!
Moved all classes into package it.radiolabs.obampxp.
I discuessed this with remo and he is fine with this change. ;)
This is needed to be able to create unit-tests in a seperate package.

build-file etc. has been update, new build is included.
If you have any problems feel free to contact me.

git-svn-id: https://svn.ninux.org/svn/ninuxdeveloping/OBAMPxP/trunk@164 ef3c9aa7-e928-0410-9a28-8491e62b674b